sessions/session-01-ai-chatbots/participants/sarah-larkin/setup/requirements.txt
google-cloud-aiplatform>=1.26.0
vertexai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0